chore: league update 페이지 컴포넌트화
